requirements.txt
django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
psycopg2==2.9.5
python-dotenv==0.21.0
requests==2.28.2